100 - Google Cloud Service Account Key/pyproject.toml
[project]
name = "gcp-sa-validation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "google-cloud-iam>=2.20.0",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-storage>=3.7.0",
    "python-dotenv>=1.2.1",
]
